# swagger.yaml
openapi: 3.0.1
info:
  title: Get it done API
  description: This API sevres for an application "Get it done", which is an tracking system of Kanban style.
  contact:
    email: lukasvrbaa@gmail.com
  license:
    name: MIT license
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: localhost:8081/api
...
paths:
...
  /teams:
    post:
      tags:
        - Teams
      summary: Create a team
      operationId: createTeam
      responses:
        200:
          description: List of users in the team.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Friendship'
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
  /teams/{teamId}/users:
    get:
      tags:
        - Teams
      summary: Get members of a team
      operationId: getTeamMembers
      parameters:
        - name: teamId
          in: path
          description: ID of a team
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of users in the team.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Friendship'
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
    put:
      tags:
        - Teams
      summary: Update team
      description: Updates team and returns it's new value
      operationId: updateTeam
      parameters:
        - name: teamId
          in: path
          description: ID of a team to be updated
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful update of a team, returns its new value
          content:
            application/json:
              schema:
                {}
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
  /teams/{teamId}:
    delete:
      tags:
        - Teams
      summary: Delete team
      description: Permanently delete a team.
      operationId: deleteTeam
      parameters:
        - name: teamId
          in: path
          description: ID of team to be deleted.
          required: true
          schema:
            type: string
      responses:
        204:
          description: Successfully deleted
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
...
  /projects:
    post:
      tags:
        - Projects
      summary: Create a project
      description: Creates a new project.
      operationId: createProject
      requestBody:
        description: Created game object
        content:
          application/json:
            schema:
              {}
      responses:
        201:
          description: Project created
          content:
            application/json:
              schema:
                type: object
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        401:
          description: User is not logged in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        403:
          description: User does not have enough rights.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
    get:
      tags:
        - Projects
      summary: Get all projects
      description: Returns all projects, that are visible for current logged in user.
      operationId: getAllProjectsForUser
      responses:
        200:
          description: Returns list of projects for current user
          content:
            application/json:
              schema:
                type: object
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
  /projects/{projectIdOrKey}:
    get:
      tags:
        - Projects
      summary: Get project description
      operationId: getOneProject
      parameters:
        - name: projectIdOrKey
          in: path
          description: ID or key of a project
          required: true
          schema:
            type: string
      responses:
        200:
          description: Project description.
          content:
            application/json:
              schema:
                {}
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'

    put:
      tags:
        - Projects
      summary: Update Project
      description: Updates team and returns it's new value
      operationId: updateProject
      parameters:
        - name: projectIdOrKey
          in: path
          description: ID of a team to be updated
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successfull update of a project.
          content:
            application/json:
              schema:
                type: object
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
    delete:
      tags:
        - Projects
      summary: Delete project
      description: Permanently delete a project
      operationId: deleteProject
      parameters:
        - name: projectIdOrKey
          in: path
          description: Project key or ID to be deleted
          required: true
          schema:
            type: string
      responses:
        204:
          description: Project successfully deleted
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrResponse'
components:
  schemas:
...
    ApiErrResponse:
      type: object
      properties:
        message:
          type: string